Move App.tsx to src folder because of many problems
