--- __tests__/App.js
@@ -1,14 +1,0 @@
-import "react-native";
-import React from "react";
-import App from "../App.tsx";
-
-// Note: test renderer must be required after react-native.
-import renderer from "react-test-renderer";
-
-it("renders correctly", () => {
-    renderer.create(<App />);
-});
-
-it("this test should succeed", () => {
-    expect(true).toBe(true);
-});
